+++ pages/dev/+Page.mdx
@@ -23,7 +23,6 @@ import { PageHeader } from "~/components";
 
 - [Concept app index](/dev/concepts)
 - [Built with Macrostrat](/dev/test-site/about)
-- [New homepage](/dev/test-site/)
 - [Documentation](/dev/docs)
 
 ## Miscellaneous
