--- pages/dev/test-site/main.module.sass
@@ -1,73 +1,0 @@
-.nav
-  position: fixed
-  top: 0
-  left: 0
-  width: 100%
-  z-index: 10000
-  display: flex
-  background-color: #015eab
-
-  a 
-    color: green
-    text-align: center
-    padding: 14px 16px
-    text-decoration: none
-
-
-.footer 
-  width: 100%
-  position: absolute
-  left: 0
-  background-color: #015EAB
-  margin: 0
-
-.footer-container
-  display: flex
-  flex-direction: row
-  justify-content: space-between
-  margin: 0 auto
-  width: 60%
-
-.col-sm-4
-  color: white
-  width: 33.333%
-  color: #E0E1E6
-  font-weight: 200
-  text-align: center
-  min-height: 1px
-  padding-left: 15px
-  padding-right: 15px
-  font-weight: bold
-  padding-bottom: 18px
-
-.nav-logo
-  padding-top: 5px
-
-.footer-nav
-  list-style-type: none
-  padding-right: 40px
-
-  li
-    padding: 5px 0
-
-  a
-    color: white
-
-  a:hover
-    text-decoration: none
-
-.git_logo
-  vertical-align: sub
-
-.f1-text
-  display: flex
-  gap: 7px
-  
-  a
-    color: white
-
-#who-made-it
-  padding-top: 25px
-
-.logo_white
-  padding-top: 20px
