--- pages/dev/test-site/main.styl
@@ -1,69 +1,0 @@
-.nav
-  position: fixed
-  top: 0
-  left: 0
-  width: 100%
-  z-index: 10000
-  display: flex
-  gap: 20px
-  background-color: #015eab;
-  height: 5vh
-  padding: 0 10px
-  align-items: center
-
-  .nav-link
-    display: flex
-    justify-content: center
-    align-items: center
-
-  a 
-    color: white !important;
-    text-align: center;
-    text-decoration: none
-
-
-.footer 
-  width: 100%
-  position: absolute
-  left: 0
-  background-color: #015EAB
-  margin: 0
-
-  .footer-nav
-    display: flex
-    flex-direction: column
-    justify-content: center
-    align-items: center
-    gap: 10px
-    
-    a
-      color: white !important
-
-    a:hover
-      text-decoration: none
-
-.footer-container
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin: 0 auto;
-  width: 60%
-
-.nav-logo
-  padding-top: 5px
-
-.git_logo
-  vertical-align: sub
-
-.f1-text
-  display: flex
-  gap: 7px
-  
-  a
-    color: white
-
-#who-made-it
-  padding-top: 25px
-
-.logo_white
-  padding-top: 20px
