+++ pages/main.module.sass
@@ -451,18 +451,3 @@ a
 
 .hide
   display: none
-
-.footer-text-container
-  display: flex
-  flex-direction: column
-  justify-content: center
-  align-items: center
-  color: white
-  padding: 20px 0
-
-.footer-text 
-  display: flex
-  gap: 2px
-
-  a
-    color: white !important
