+++ pages/people/+Page.client.ts
