+++ pages/people/main.styl
